feat(9.2): handle settings (#221)
