+++ src/antares/craft/model/settings/adequacy_patch.py
@@ -23,13 +23,14 @@ class PriceTakingOrder(Enum):
 class AdequacyPatchParameters:
     include_adq_patch: bool = False
     set_to_null_ntc_from_physical_out_to_physical_in_for_first_step: bool = True
-    set_to_null_ntc_between_physical_out_for_first_step: bool = True
     price_taking_order: PriceTakingOrder = PriceTakingOrder.DENS
     include_hurdle_cost_csr: bool = False
     check_csr_cost_function: bool = False
     threshold_initiate_curtailment_sharing_rule: int = 1
     threshold_display_local_matching_rule_violations: int = 0
     threshold_csr_variable_bounds_relaxation: int = 7
+    # Parameters removed since v9.2
+    set_to_null_ntc_between_physical_out_for_first_step: Optional[bool] = None  # was True in v8.8
 
 
 @dataclass
