+++ src/antares/craft/model/study.py
@@ -112,7 +112,7 @@ class Study:
         self._settings = self._settings_service.read_study_settings()
 
     def update_settings(self, settings: StudySettingsUpdate) -> None:
-        self._settings_service.edit_study_settings(settings)
+        self._settings_service.edit_study_settings(settings, self._version)
         new_settings = self._settings_service.read_study_settings()
         self._settings.general_parameters = new_settings.general_parameters
         self._settings.optimization_parameters = new_settings.optimization_parameters
